requirements.txt
Flask>=3.0.0
google-cloud-bigquery>=3.13.0
google-cloud-aiplatform>=1.38.0
pandas>=2.1.0
python-dotenv>=1.0.0
gunicorn>=21.2.0